Supports composer through a bin script
Installing `composer` with the provided `php` script is not possible and
making it available poses some other problems like mapping a directory
available in the $PATH and putting that script there.

By providing a `composer` executable, one can have an environment with
everything needed to develop PHP applications. I only tested this on
Linux but OSX should be fine - who knows how this works under Windows.
Some niceties are provided:

* SSH socket is mapped so keys can be exchanged for private repositories
* COMPOSER_HOME environment is always used
* Execution is always made with the current user
* Current directory is always mapped to /app so project dependencies
  behave as expected
* Installing dependencies ignore platform requirements are the PHP
  version being used is always the lates available